+++ app/Http/Requests/UpdateJobRequest.php
@@ -3,13 +3,16 @@
 namespace App\Http\Requests;
 
 use App\Job;
+use App\User;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Foundation\Http\FormRequest;
 
 class UpdateJobRequest extends FormRequest
 {
     public function authorize()
     {
-       // return \Gate::allows('job_edit');
+        $user = User::find(Auth::user()->id);
+        return $user-> hasRole('client');
     }
 
     public function rules()