+++ app/Http/Requests/UpdateProposalRequest.php
@@ -2,14 +2,17 @@
 
 namespace App\Http\Requests;
 
+use App\User;
 use App\Proposal;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Foundation\Http\FormRequest;
 
 class UpdateProposalRequest extends FormRequest
 {
     public function authorize()
     {
-       // return \Gate::allows('proposal_edit');
+        $user = User::find(Auth::user()->id);
+        return $user-> hasRole('freelancer');
     }
 
     public function rules()