+++ changelog.md
@@ -1,5 +1,5 @@
 - ## Commit 57th:
-     ### Major Controller, Model and Database Changes:
+     ### **Major Controller, Model and Database Changes:**
 
     1. Database Proposal and Job new Column _(user_id & references)_.
 
@@ -11,4 +11,17 @@
 
     5. Unnecessary Clutters Deleted.
 
-    6. 
+<br><hr><br>
+
+- ## Commit 59th:
+     ### **New Project Post and Eloquent Views Added:**
+
+    1. Jobs can be posted by **Client**.
+
+    2. **Clients** can view their posted jobs in (my-jobs) page.
+
+    3. Proposals _Count_ and _Project Tracking_ implemented.
+
+    4. New Frontend Pages Added.
+
+    5. `Controllers` & `Models` many changes.